requirements.txt
streamlit==1.30.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.3
pyyaml==6.0.1
plotly==5.18.0
requests==2.31.0
statsmodels==0.14.0
prophet==1.1.5
xgboost==2.0.3
lightgbm==4.1.0
catboost==1.2.2
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0
torch==2.2.0
matplotlib==3.8.0
seaborn==0.13.0
